Introduce MixIns file and ass DBEnumMixIn
