+++ rotkehlchen/chain/ethereum/transactions.py
@@ -9,7 +9,7 @@ from rotkehlchen.externalapis.etherscan import Etherscan
 from rotkehlchen.logging import RotkehlchenLogsAdapter
 from rotkehlchen.typing import ChecksumEthAddress, EthereumTransaction, Timestamp
 from rotkehlchen.user_messages import MessagesAggregator
-from rotkehlchen.utils.interfaces import LockableQueryObject, protect_with_lock
+from rotkehlchen.utils.mixins import LockableQueryMixIn, protect_with_lock
 
 logger = logging.getLogger(__name__)
 log = RotkehlchenLogsAdapter(logger)
@@ -17,7 +17,7 @@ log = RotkehlchenLogsAdapter(logger)
 FREE_ETH_TX_LIMIT = 250
 
 
-class EthTransactions(LockableQueryObject):
+class EthTransactions(LockableQueryMixIn):
 
     def __init__(
             self,
