+++ rotkehlchen/exchanges/poloniex.py
@@ -56,8 +56,8 @@ from rotkehlchen.typing import (
     TradePair,
 )
 from rotkehlchen.user_messages import MessagesAggregator
-from rotkehlchen.utils.interfaces import cache_response_timewise, protect_with_lock
 from rotkehlchen.utils.misc import create_timestamp, ts_now_in_ms
+from rotkehlchen.utils.mixins import cache_response_timewise, protect_with_lock
 from rotkehlchen.utils.serialization import jsonloads_dict, jsonloads_list
 
 if TYPE_CHECKING:
